requirements.txt
aws_cdk.aws_directoryservice
aws_cdk.aws_ec2
aws_cdk.aws_iam
aws_cdk.core
